requirements.txt
langchain
langchain_community
openai
together
python-dotenv
gradio